requirements.txt
python-telegram-bot
requests
discord.py